requirements.txt
streamlit
numpy
pandas
matplotlib
seaborn
Pillow
opencv-python
tensorflow
scikit-learn
google-generativeai
sqlite3
fpdf
PyYAML
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0